% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/iHeatmap.R
\name{iHeatmap}
\alias{iHeatmap}
\title{Interactive Heatmap}
\source{
The interface was designed based on Joe cheng's \link{d3heatmap}.
}
\usage{
iHeatmap(x, colAnnote = NULL, rowAnnote = NULL, width = NULL,
  height = NULL, colors = "RdYlBu", ClustM = "complete",
  distM = "euclidean", Colv = TRUE, Rowv = TRUE, xaxis_height = 100,
  yaxis_width = 100, anim_duration = 500, showHeat = TRUE,
  addOnInfo = NULL, scale = FALSE, col_scale = TRUE,
  cor_method = "pearson", font_size = 10, probs = 100, annote_pad = 7,
  legend_width = 50, ...)
}
\arguments{
\item{x}{A numeric \code{m x n} matrix.}

\item{colAnnote}{Column annotations-
Takes a \code{n x 1} matrix.}

\item{rowAnnote}{Row annotations-
Takes a \code{m x 1} matrix.}

\item{width}{Width in pixels (optional, defaults to automatic sizing).}

\item{height}{Height in pixels (optional, defaults to automatic sizing).}

\item{colors}{Either a colorbrewer2.org palette name (e.g. \code{"RdYlBu"} or
\code{"Blues"}).}

\item{ClustM}{The agglomeration (Linkage) method to be used.
This should be (an unambiguous abbreviation of) one of \code{"ward.D"}, \code{"ward.D2"},
\code{"single"}, \code{"complete"}, \code{"average"} (= UPGMA), \code{"mcquitty"} (= WPGMA),
\code{"median"} (= WPGMC) or \code{"centroid"} (= UPGMC).}

\item{distM}{The distance measure to be used. This must be one of \code{"correlation"},\code{"euclidean"},
\code{"maximum"}, \code{"manhattan"}, \code{"canberra"}, \code{"binary"} or \code{"minkowski"}.}

\item{Colv}{Determines if columns should be clustered (defaults to \code{TRUE})}

\item{Rowv}{Determines if rows should be clustered (defaults to \code{TRUE})}

\item{showHeat}{Only show the dendrogram.}

\item{addOnInfo}{For adding on extra information}

\item{scale}{Determine if the data is to be scaled, defaults to \code{TRUE}}

\item{col_scale}{Determins if data is to be scaled by row or column, defaults to \code{TRUE} (scale by column)}

\item{cor_method}{Determins correlation method, defaults to \code{"pearson"}}

\item{probs}{Determines the quantile, defaults to \code{100}}
}
\description{
Creates a D3.js-based heatmap widget.
}
\examples{
library(iHeatmap)
iHeatmap(mtcars, scale = TRUE, col_scale=TRUE, colors = "Blues")
}

