% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iHeatmap.R
\name{iHeatmapOutput}
\alias{iHeatmapOutput}
\alias{renderIHeatmap}
\title{Wrapper functions for using d3heatmap in shiny}
\usage{
iHeatmapOutput(outputId, width = "100\%", height = "400px")

renderIHeatmap(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{outputId}{Output variable to read from}

\item{width, height}{The width and height of the map (see
\link[htmlwidgets]{shinyWidgetOutput})}

\item{expr}{An expression that generates a \code{\link{iHeatmap}} object}

\item{env}{The environment in which to evaluate \code{expr}}

\item{quoted}{Is \code{expr} a quoted expression (with \code{quote()})? This
is useful if you want to save an expression in a variable.}
}
\description{
Use \code{iHeatmapOutput} to create a UI element, and \code{renderIHeatmap}
to render the heatmap.
}
\examples{
\donttest{
library(iHeatmap)
library(shiny)

ui <- fluidPage(
headerPanel("A heatmap demo"),
sidebarPanel(
  selectInput("palette", "Palette", c("YlOrRd", "RdYlBu", "Greens", "Blues")),
  selectInput("cluster_method", "Clustering Method", c('complete', 'average','ward.D2','single','mcquitty','median','centroid')),
  selectInput("cluster_row","Cluster rows",c('TRUE', 'FALSE')),
  selectInput("cluster_col","Cluster columns",c('TRUE','FALSE')),
  selectInput("dist","Distance method",c('euclidean','correlation','maximum','manhattan','canberra','binary','minkowski'))
),
mainPanel(
  iHeatmapOutput("heatmap")
)
)

server <- function(input, output, session) {
  output$heatmap <- renderIHeatmap({
    col_annot <- matrix(runif(11),11,1)
    row_annot <- matrix(runif(32),32,1)
    iHeatmap(
      mtcars,colAnnote=round(col_annot,1),rowAnnote= round(row_annot,1),
      scale="column",colors = input$palette,
      ClustM= input$cluster_method, distM= input$dist,
      Colv = input$cluster_col,Rowv = input$cluster_row
    )
  })
}

shinyApp(ui, server)
}

}

